refactor: merge cue editor into sound studio as Cues tab
Sound cue editor is now a tab of the sound studio page — no separate
URL or folder picker needed. Opening shared/assets/ loads both sounds
and cues in one step. Navigating between tabs preserves all state.

- CueEditor.tsx: extracted cue editor as a self-contained component
- page.tsx: adds Cues tab, passes cueFiles + parsedBin to CueEditor
- cues/page.tsx: now redirects to /sound-studio?tab=cues
- Tab state driven by ?tab= URL param for deep-linking
- useEffect persists store on state changes (fixes navigation loss)
